use float32

+++ demo_rotary.py
@@ -99,7 +99,7 @@ def check_rope():
 
 def run_llm():
     from vllm import LLM, SamplingParams
-    llm = LLM(model="THUDM/chatglm3-6b", trust_remote_code=True, dtype='float32')
+    llm = LLM(model="THUDM/chatglm3-6b", trust_remote_code=True)
     # print(len(llm.llm_engine.workers), llm.llm_engine.workers[0].model)
     prompts = [
         "Hello, my name is",
